+++ priv/repo/migrations/20220526160059_add_pending_removal_to_grants.exs
@@ -1,0 +1,9 @@
+defmodule TdDd.Repo.Migrations.AddPendingRemovalToGrants do
+  use Ecto.Migration
+
+  def change do
+    alter table("grants") do
+      add :pending_removal, :boolean, null: false, default: false
+    end
+  end
+end